requirements_dev.txt
pre-commit~=4.3; python_version>="3.9"
pylint~=3.3; python_version>="3.9"
pytest-cov~=6.2; python_version>="3.9"
